requirements.txt
fastapi==0.115.6
prisma==0.15.0
pydantic==2.10.4
pydantic_core==2.27.2
uvicorn==0.34.0
python-socketio==5.11.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
redis==5.1.1
shortuuid==1.0.13
uuid6==2024.1.12



